requirements.txt
torch>=2.0
transformers>=4.21
accelerate
datasets>=1.8
biopython>=1.80
fair-esm
sentencepiece
tqdm
numpy
PyYAML
